update(cmd): setup the CLI configuration flags and their validations
The flags take precedence over environment variables.
The environment variables have a prefix (`LSTN_`) and they take precedence of the configuration file.
The configuration file is a `.lstn.yaml` file provided by the `--config` flag or existing in the home directory.

The configuration options get passed through the command context.

Signed-off-by: Leonardo Di Donato <leodidonato@gmail.com>