chore: golangci-lint format print affected file
Q: why are we changing the output format?
A: because the current output format is not informative about where the
issue is, it doesn't show file names for the given report, for example.
Which is very frustrating. The `tab` output format does that.

Q: what does the `IGNORE_FUTURE_WARNINGS` do?
A: it's there so golangci-lint does not complain we change the output
format